Refine basic_guidance.prompt (negation removal)
Reworded the parts of this prompt which involve negation. Sometimes LLMs struggle with negated sentences and can interpret negative instructions as the *opposite* of what you think you're asking it to do. So I've replaced the negative sentences with positive counterparts, focusing on what it should actually be doing (instead of only describing what it's not supposed to do).

This should improve the consistency||reliability of generated output.