Add pagination support for output
In some cases, it's useful to limit long tool output*. This feature adds pagination support.

It adds two command line parameters to `run` and `serve`, `--paginate` and `--max-output`. Using `--paginate` will enable the feature. `--max-output` is set to 4096 by default.

Enabling paginate will have an optional `page` parameter added to all tool functions. It will also truncate the output to `max-output` size, and provide hints back to the LLM that there are other pages.

Since the `page` parameter is optional, if the tool is run as normal with `paginate` enabled then the output will only be truncated if it exceeds `max-output` otherwise it will behave as normal with no change.

The change is made by injecting the optional `page` parameter via the jinja template. The rest of the code is mostly making sure that the command line option is propagated all the way down to the templating. I'm a little uncertain that I did this is the right way, and would be open to direction if there's a better way.

* I've found two specific use cases for limiting output:

1) Claude Desktop implements three separate limits for tool response (1M), message size (10k characters) and a varying conversation length.

2) Reducing input tokens for cost purposes.

Often hints from the first few pages are enough for very long tool output e.g. to get the LLM to generate more targeted searches for the content it needs. Or knowing the content is at the end of a long output, the last page could be grabbed without the rest of the output.